feat: copy with tree .

+++ internal/tree/red_black_tree.go
@@ -30,7 +30,7 @@ const (
 //  4. every path from root to leaf node has the same number of black nodes
 type RBTree[K any, V any] struct {
 	root *rbNode[K, V]
-	size int64
+	size int
 	cmp  easy_kit.Comparator[K]
 }
 
@@ -42,7 +42,7 @@ func NewRBTree[K any, V any](cmp easy_kit.Comparator[K]) *RBTree[K, V] {
 	}
 }
 
-func (rbt *RBTree[K, V]) Size() int64 {
+func (rbt *RBTree[K, V]) Size() int {
 	return rbt.size
 }
